refactor(vroce): move helpers and consts
to appropriate location.
Make the component cleaner and readable.
