+++ code/ali-je-vroce/vroce.jsx
@@ -2,226 +2,9 @@ import { createSignal, Show, For, onMount, createEffect } from "solid-js";
 import { Select } from "@kobalte/core/select";
 import { IsItHotDot } from "../components/is-it-hot-dot.jsx";
 
-/**
- * @typedef {'p00' | 'p05' | 'p20' | 'p40' | 'p60' | 'p80' | 'p95'} ValidPercentile
- * @typedef { ValidPercentile | ''} PercentileKey
- */
-
-const DEFAULT_STATION_ID = 1495; // Ljubljana
-const DEFAULT_STATION_NAME = 'Ljubljana';
-const DEFAULT_STATION_PREFIX = 'v'; // 'v' for "v Ljubljani"    
-
-const DEFAULT_STATION = {
-    value: DEFAULT_STATION_ID,
-    label: DEFAULT_STATION_NAME,
-    prefix: DEFAULT_STATION_PREFIX,
-}
-
-// URL
-const baseUrl = 'https://stage-data.podnebnik.org'
-// const baseUrl = 'http://localhost:8010'
-const vremenarBaseUrl = 'https://podnebnik.vremenar.app/staging'
-// const vremenarBaseUrl = 'http://localhost:8000'
-
-/**
- * @type {Record<PercentileKey, string>} PercentileLabels
- */
-const percentile_labels = {
-    'p00': 'manj kot 5 %',
-    'p05': '5 %',
-    'p20': '20 %',
-    'p40': '40 %',
-    'p60': '60 %',
-    'p80': '80 %',
-    'p95': '95 %',
-    '': '',
-}
-
-/**
- * @type {Record<PercentileKey, string>} PercentileValues
- */
-const vrednosti = {
-    'p00': 'Niti pod razno',
-    'p05': 'Ne!',
-    'p20': 'Ne.',
-    'p40': 'Niti ne.',
-    'p60': 'Ja.',
-    'p80': 'Ja!',
-    'p95': 'Ja, absolutno!',
-    '': '',
-}
-
-/**
- * @type {Record<PercentileKey, string>} PercentileDescriptions
- */
-const opisi = {
-    'p00': 'Se hecaš?! Ful je mraz!',
-    'p05': 'Pravzaprav je res mrzlo.',
-    'p20': 'Dejansko je kar hladno.',
-    'p40': 'Precej povprečno.',
-    'p60': 'Je topleje kot običajno.',
-    'p80': 'Res je vroče!',
-    'p95': 'Peklensko vroče je!',
-    '': '',
-}
-
-/**
- * Returns a string representation of a number, prefixing it with a zero if it is less than 10.
- *
- * @param {number} number - The number to format.
- * @returns {string} The zero-prefixed number as a string if less than 10, otherwise the number as a string.
- */
-function zeroPrefix(number) {
-    if (number < 10) {
-        return `0${number}`
-    }
-    return `${number}`
-}
-
-/**
- * Formats a JavaScript Date object into a string suitable for queries (YYYY-MM-DD).
- *
- * @param {Date} date - The date to format.
- * @returns {string} The formatted date string in 'YYYY-MM-DD' format.
- */
-function formatDateForQuery(date) {
-    let year = `${date.getFullYear()}`
-    let month = `${zeroPrefix(date.getMonth() + 1)}`
-    let day = `${zeroPrefix(date.getDate())}`
-    return `${year}-${month}-${day}`
-}
-
-/**
- * Formats a given date as a human-readable string indicating whether the date is today or yesterday,
- * and includes the time in HH:MM format.
- *
- * @param {Date} date - The date to format.
- * @param {Date} updated - The reference date to compare against for determining "today" or "yesterday".
- * @returns {string} A formatted string such as "danes ob 14:30" or "včeraj ob 09:15".
- */
-function formatTime(date, updated) {
-    let day = date.getDate() == updated.getDate() ? 'danes' : 'včeraj'
-    let time = `${zeroPrefix(date.getHours())}:${zeroPrefix(date.getMinutes())}`
-
-    return `${day} ob ${time}`
-}
-
-/**
- * Asynchronously loads a list of temperature stations from a remote JSON endpoint.
- *
- * Fetches station data, processes each station row, and returns an array of station objects.
- * Each object contains the station's ID, the locative name (excluding the prefix), and the prefix itself.
- *
- * @async
- * @function loadStations
- * @returns {Promise<Array<{station_id: string, name_locative: string, prefix: string}>>}
- *   Resolves to an array of station objects, or an empty array if the fetch fails.
- */
-async function loadStations() {
-    const resultStations = await fetch(`${baseUrl}/temperature/temperature~2Eslovenia_stations.json?&_col=station_id&_col=name&_col=name_locative&_sort=name`);
-    if (resultStations.ok) {
-        let stationsList = [];
-        const dataStations = await resultStations.json();
-        for (let row of dataStations["rows"]) {
-            let name_list = row[3].split(' ')
-            stationsList.push({
-                'station_id': row[1],
-                'name_locative': name_list.slice(1).join(" "),
-                'prefix': name_list[0],
-            })
-        }
-        return stationsList;
-    }
-    console.error('Failed to load stations');
-    return [];
-}
-
-/**
- * Fetches and processes temperature data for a given weather station.
- *
- * Retrieves the latest average, minimum, and maximum temperatures, as well as historical percentile data,
- * for the specified station. Determines the percentile bucket for the current average temperature.
- *
- * @async
- * @param {string} stationID - The ID of the weather station to fetch data for.
- * @returns {Promise<{
- *   resultValue: string,
- *   resultTemperatureValue: number|string,
- *   tempMin: number|string,
- *   timeMin: string,
- *   tempMax: number|string,
- *   timeMax: string,
- *   tempAvg: number|string,
- *   timeUpdated: string
- * }>} An object containing temperature statistics and percentile results. If the request fails, returns empty strings for all fields.
- */
-async function requestData(stationID) {
-    const resultAverage = await fetch(`${vremenarBaseUrl}/stations/details/METEO-${stationID}?country=si`);
-    if (resultAverage.ok) {
-        const dataAverage = await resultAverage.json();
-        const averageTemperature = dataAverage.statistics.temperature_average_24h;
-
-        let timeUpdated = new Date(Number(dataAverage.statistics.timestamp))
-
-        const date = formatDateForQuery(timeUpdated)
-
-        let timeMinDate = new Date(Number(dataAverage.statistics.timestamp_temperature_min_24h))
-        let timeMaxDate = new Date(Number(dataAverage.statistics.timestamp_temperature_max_24h))
-
-        const resultPercentile = await fetch(`${baseUrl}/temperature/temperature~2Eslovenia_historical~2Edaily~2Eaverage_percentiles.json?date__exact=${date}&station_id__exact=${stationID}&_col=p05&_col=p20&_col=p40&_col=p60&_col=p80&_col=p95`);
-        if (resultPercentile.ok) {
-            const dataPercentile = await resultPercentile.json();
-
-            let columns = dataPercentile['columns'];
-            let values = dataPercentile['rows'][0];
-            columns.shift();
-            values.shift();
-
-            let resultValue = -1
-            let resultTemperatureValue = -1
-            if (averageTemperature < values[0]) {
-                resultValue = 'p00'
-                resultTemperatureValue = values[0]
-            } else {
-                for (let i = 0; i < values.length; i++) {
-                    if (i == values.length - 1) {
-                        resultValue = 'p95'
-                        resultTemperatureValue = values[i]
-                    } else if (averageTemperature >= values[i] && averageTemperature < values[i + 1]) {
-                        resultValue = columns[i]
-                        resultTemperatureValue = values[i]
-                        break
-                    }
-                }
-            }
-
-            return {
-                resultValue,
-                resultTemperatureValue,
-                tempMin: dataAverage.statistics.temperature_min_24h,
-                timeMin: formatTime(timeMinDate, timeUpdated),
-                tempMax: dataAverage.statistics.temperature_max_24h,
-                timeMax: formatTime(timeMaxDate, timeUpdated),
-                tempAvg: averageTemperature,
-                timeUpdated: new Intl.DateTimeFormat('sl-SI', {
-                    dateStyle: 'long',
-                    timeStyle: 'short',
-                }).format(timeUpdated),
-            }
-        }
-    }
-    console.error(`Failed to load data for station ${stationID}`);
-    return {
-        resultValue: '',
-        resultTemperatureValue: '',
-        tempMin: '',
-        timeMin: '',
-        tempMax: '',
-        timeMax: '',
-        tempAvg: '',
-        timeUpdated: '',
-    };
-}
+import { requestData, loadStations } from "./helpers.mjs";
+import { vrednosti, opisi, percentile_labels } from "./constants.mjs";
+import { DEFAULT_STATION } from "./constants.mjs";
 
 /**
  * AliJeVroce is a Solid JS component that displays whether it is hot today in a selected location,
@@ -326,7 +109,7 @@ export function AliJeVroce() {
         </p>
         <p></p>
         <p class="font-black text-6xl">
-            <IsItHotDot color={result() === "" ? "initial" : result()} class="mr-8" />{" "}
+            <IsItHotDot color={result()} class="mr-8" />{" "}
             {vrednosti[result()]}
         </p>
         <p class="text-4xl font-semibold">{opisi[result()]}</p>
